revert: temporarily modify CodeQL queries

+++ numeric/source/prime.cpp
@@ -54,7 +54,7 @@ std::vector<std::uint32_t> Prime::euler(const std::uint32_t max)
 
         for (std::uint32_t j = 1; (j <= container.size()) && ((i * container[j - 1]) <= max); ++j)
         {
-            isPrime[i * container[j - 1]] = false;
+            isPrime[static_cast<std::uint32_t>(i * container[j - 1])] = false;
             if (0 == (i % container[j - 1]))
             {
                 break;
